+++ dockers/release-conda/conda_entrypoint.sh
@@ -1,5 +1,0 @@
-#!/bin/bash --login
-set -e
-LD_LIBRARY_PATH=/usr/local/nvidia/lib:/usr/local/nvidia/lib64
-. ~/.bashrc
-exec "$@"