+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [2.6.0] - 2024-XX-XX
+
+### Added
+
+- Support for Lightning and PyTorch ``2.6.0``
+
+### Deprecated
+
+- removed support for PyTorch `2.2`
+- removed use of conda builds (aligning with upstream PyTorch)
+
 ## [2.5.0] - 2024-XX-XX
 
 ### Added